remove TransformedColumn
