+++ snuba/query/conditions.py
@@ -2,9 +2,27 @@ from typing import Mapping, Optional, Sequence
 
 from snuba.query.dsl import literals_tuple
 from snuba.query.expressions import Expression, FunctionCall, Literal
-from snuba.query.functions import BooleanFunctions, ConditionFunctions
 from snuba.query.matchers import FunctionCall as FunctionCallPattern
 from snuba.query.matchers import AnyExpression, Or, Param, Pattern, String
+
+
+class ConditionFunctions:
+    """
+    Function names for comparison operations.
+    """
+
+    EQ = "equals"
+    NEQ = "notEquals"
+    LTE = "lessOrEquals"
+    GTE = "greaterOrEquals"
+    LT = "less"
+    GT = "greater"
+    IS_NULL = "isNull"
+    IS_NOT_NULL = "isNotNull"
+    LIKE = "like"
+    NOT_LIKE = "notLike"
+    IN = "in"
+    NOT_IN = "notIn"
 
 
 OPERATOR_TO_FUNCTION: Mapping[str, str] = {
@@ -26,6 +44,16 @@ OPERATOR_TO_FUNCTION: Mapping[str, str] = {
 FUNCTION_TO_OPERATOR: Mapping[str, str] = {
     func: op for op, func in OPERATOR_TO_FUNCTION.items()
 }
+
+
+class BooleanFunctions:
+    """
+    Same as comparison functions but for boolean operators.
+    """
+
+    NOT = "not"
+    AND = "and"
+    OR = "or"
 
 
 UNARY_OPERATORS = [
