--- snuba/query/functions.py
@@ -1,27 +1,0 @@
-class ConditionFunctions:
-    """
-    Function names for comparison operations.
-    """
-
-    EQ = "equals"
-    NEQ = "notEquals"
-    LTE = "lessOrEquals"
-    GTE = "greaterOrEquals"
-    LT = "less"
-    GT = "greater"
-    IS_NULL = "isNull"
-    IS_NOT_NULL = "isNotNull"
-    LIKE = "like"
-    NOT_LIKE = "notLike"
-    IN = "in"
-    NOT_IN = "notIn"
-
-
-class BooleanFunctions:
-    """
-    Same as comparison functions but for boolean operators.
-    """
-
-    NOT = "not"
-    AND = "and"
-    OR = "or"
