+++ snuba/query/matchers.py
@@ -11,7 +11,6 @@ from snuba.query.expressions import Column as ColumnExpr
 from snuba.query.expressions import Expression, OptionalScalarType
 from snuba.query.expressions import FunctionCall as FunctionCallExpr
 from snuba.query.expressions import Literal as LiteralExpr
-from snuba.query.functions import BooleanFunctions, ConditionFunctions
 
 MatchType = Union[Expression, OptionalScalarType]
 
@@ -322,45 +321,6 @@ class FunctionCall(Pattern[FunctionCallExpr]):
                     result = result.merge(p_result)
 
         return result
-
-
-@dataclass(frozen=True)
-class TransformedColumn(Pattern[FunctionCallExpr]):
-    """
-    Matches a Column that has been wrapped in zero or more Functions. It recursively
-    checks the parameters of the FunctionCall to see if at least one of them is a Column
-    matching the provided column name.
-
-    Note that this makes no guarantees about the functions being called, it naively
-    goes through all the parameters and returns the first column match it finds.
-    """
-
-    column: Column
-
-    def match(self, node: AnyType) -> Optional[MatchResult]:
-        if not isinstance(node, (ColumnExpr, FunctionCallExpr)):
-            return None
-
-        if isinstance(node, ColumnExpr):
-            return self.column.match(node)
-
-        if isinstance(node, FunctionCallExpr):
-            if len(node.parameters) == 0:
-                return None
-
-            # Matching on these is dangerous since they are used for conditions. It's possible to match
-            # the right or left side of a condition, but not the condition as a whole.
-            if hasattr(BooleanFunctions, node.function_name) or hasattr(
-                ConditionFunctions, node.function_name
-            ):
-                return None
-
-            for parameter in node.parameters:
-                partial_result = self.match(parameter)
-                if partial_result is not None:
-                    return partial_result
-
-        return None
 
 
 # TODO: Add more Patterns when needed.
